+++ tests/test_textrenderer.py
@@ -1,0 +1,14 @@
+from qutip_qip.circuit import QubitCircuit
+from qutip_qip.circuit.text_renderer import TextRenderer
+
+
+def test_parts_lenght():
+    """
+    Check if the lenght of different parts of the gate is same
+    """
+
+    # render ISWAP gate
+    qc = QubitCircuit(2)
+    
+    
+
